refactor `romanize` method
To handle the case when the ` torch.argmax` raise an exception is that
the target_tensor_logits return an empty tensor torch.Size([0]).
This due to the following code segment within the
method named `forward` of Seq2seq class

```
 for di in range(max_len):
     ...
     if inference and decoder_input == end_token:
                return outputs[:di]
    ...
```

if the variable `di` is 0 (first iteration), and the first decoder output
is  the end token <END>, it would return outputs[:0] which is an
empty Tensor.