Update augment.rst
The enhanced documentation for the pythainlp.augment module brings about several notable improvements. These changes focus on providing users with a more comprehensive understanding of the module and its various components for text augmentation in the Thai language. Here's an overview of the key changes:

Introduction: The documentation now starts with a clear introduction, emphasizing the importance of text augmentation in NLP and its specific relevance to the Thai language. This introduction sets the stage for the entire module, making it clear why text augmentation is a crucial task.

TextAugment Class: The central TextAugment class is highlighted, and its purpose as the core component of the module is explained. Users can now understand that this class serves as the gateway to various text augmentation techniques.

Class Details: Each class within the module, such as WordNetAug, Word2VecAug, FastTextAug, and BPEmbAug, is provided with a detailed description of its purpose and capabilities. This clarity allows users to determine which class is best suited for their specific text augmentation needs.

Function Descriptions: The postype2wordnet function's role in mapping part-of-speech tags to WordNet-compatible POS tags is clearly explained, facilitating the integration of WordNet augmentation with Thai text. Users can better understand how to work with this function in their text augmentation tasks.

Usage Guidance: The documentation emphasizes that users can refer to the official PyThaiNLP documentation for detailed usage examples and guidelines. This encourages users to explore the module's full potential for enriching and diversifying Thai text data and improving NLP models and applications.

These changes make the documentation more informative and accessible, making it easier for researchers, developers, and practitioners to understand how to leverage the pythainlp.augment module effectively. With this enhanced documentation, users can confidently harness the power of text augmentation for Thai language NLP tasks.