Update generate.rst
Introduction and Purpose:

The documentation for the pythainlp.generate module has been improved to offer a more explicit introduction. It now clearly defines the purpose of this module, emphasizing its role in Thai text generation within PyThaiNLP. This ensures that users have a solid understanding of what this module is designed for.
Individual Class and Function Explanations:

Each class and function within the module is explained in detail. The purpose and usage of the Unigram, Bigram, and Trigram classes, as well as the pythainlp.generate.thai2fit.gen_sentence function, and the WangChanGLM class, are highlighted. Users can now understand which language models they can use and how to choose the right one for their text generation needs.
Usage Guidelines:

A new "Usage" section is included, outlining clear steps for users on how to make use of the text generation capabilities offered by the module. These steps simplify the process and provide a structured approach to generating text.

Example:
A practical usage example is provided, demonstrating how to generate text using the Unigram class. This example gives users a reference point for applying the module in their own projects, making it more accessible.
Overall Structure and Clarity:

The documentation maintains a consistent structure with clear headings, subheadings, and bullet points, enhancing its readability and ease of navigation.