Update spell.rst
Introduction and Purpose:

The documentation for the pythainlp.spell module has undergone significant improvements. It now provides a more explicit and detailed introduction, emphasizing the module's importance in enhancing text accuracy through spelling correction. Users are made aware that it offers a range of functionalities for spell-checking and correction in the Thai language.
Function Descriptions:

Each function within the module is described in detail, outlining its specific purpose and how it can be used. Users can now understand the functionalities of correct, correct_sent, spell, and spell_sent in both single-word and sentence-level contexts.
NorvigSpellChecker Class:

The NorvigSpellChecker class is introduced as a core component of the pythainlp.spell module. Users can now understand its significance in implementing spell-checking algorithms and its potential for advanced spell-checking with customizable settings.
DEFAULT_SPELL_CHECKER:

The DEFAULT_SPELL_CHECKER instance, pre-configured with the standard NorvigSpellChecker settings and Thai National Corpus data, is presented. Users can grasp the idea of a reliable default spell-checking configuration for common use cases.
References:

The documentation now includes a "References" section, providing a citation and a link to Peter Norvig's influential work on spelling correction. This adds credibility and gives users the option to explore the academic source for more in-depth understanding.